fix warnings and errors found with -Wall -Wextra
 - configured with CPPFLAGS="-Wall -Wextra", built, and fixed all warnings
 - fixed code paths that resulted in use of uninitialized variables
 - fixed unchecked comparisons between signed and unsigned vars
 - marked unused params
 - moved static functions out of header files
 - removed useless "const" modifier on integral return values

+++ src/Interface/TextList.cpp
@@ -417,7 +417,7 @@ void TextList::setAlign(TextHAlign align, int col)
 {
 	if (col == -1)
 	{
-		for (int i = 0; i <= _columns.size() - 1; ++i)
+		for (size_t i = 0; i <= _columns.size() - 1; ++i)
 		{
 			_align[i] = align;
 		}
